Refined the Switch Object System, Have Climbables on Moving Platforms, Establish the Waterfall Riding Mechanic
Create an object called the Switch Object System (SOS) that houses the following objects as children:
A Switch that sets the SOS from false to true
An object that has two states (Ex: a moving platform that moves from one position to another depending on it's boolean, or an object that is hazardous to touch and then turns off it's hazard collider depending on it's boolean/ or a platform that makes a Swingable slide up/down/ out of it's slot depending on the boolean, or a switch that moves a water area up or down (change the water level))
Create an Object (Switch) that flips the SOS's switch state from false to true when hit by a  "Player" or "Bullet" object
Give the Switch a isOneWay boolean. If it is true, then hitting the switch once only flips the SOS once  - if(isOneWay && !SOS.SwitchState) { SOS.SwitchState = true;  collider.setActive = false} else {SOS.SwitchState = !SOS.SwitchState; collider.setActive = false}
Have the Switch have a collider that turns off .
have the switch have a float timer that turns the collider on again after a set time. It starts counting down from its max once the OnTriggerEnter2D function for the switch is activated by a "Player" or "Bullet" collider

Make sure that the player moves along with the ladder by making the player a child of the ladder object when attached
Make larger terrain with the Ladder Property, so that Flow can climb up mountains and stuff
Make Climbable Surfaces on Moving Platforms.
Have Floe be able to move on those Moving Surfaces
Have Floe get stuck on walls if a moving climbable surface passes through and floe is in the way
Have flow get off climbable surfaces, both moving or stationary, when grounded

By pressing up on a waterfall, the player will become intangible and climb up the waterfall at high-speeds, popping up when they reach the top