StageClear UI Implemented, Timer Increments now instead of counting down, Bullet hits Blade bug corrected, Can load up the stage as a prefab, Adjusted prefabs and Game Objects
Timer Seconds includes the 60. Fix that - DONE -

Timer can count up or down depending on if TimerCount is true or false - DONE -
Hitting the goal doesn't make the gameplay manager register the stage as done - Fixed -
When a bullet hits a blade, the bullet count of the player refreshes correctly -DONE-
StageClear UI is made and can slide in and show scores -DONE-
The stage is stored as a prefab with its own Stage Class. Gameplay Manager can load the stage and it's checkpoints to track. -DONE-