Created EnemyBatch System, Started work on getting stage transitions to work
Created the EnemyBatch System. When all enemies are defeated, the system activates, moving the walls away.
Created prefabs for the Player again and for the Gameplay and Transition Managers
Enemies cannot trigger switches, censors, or automove triggers
Adjusted the Prefab for Midair Soldier
Try removing the Ammo Limit
You're able to have the GameplayManager call the TransitionManager to go to a new stage
But actually loading the stage doesn't quite work. There are a few hiccups.