Fixed Stage Transition and other Bug Fixes
Fix the transition from one stage to the next. Keep in mind editing code to replicate the issue if it doesn't happen naturally -DONE-
Fixed the Health Display. It accurately represents the amount of health you have, regardless of what the current and max health are. - DONE-
Readjusted Switch code so that Bullets can hit it again. You checked only if it were the player touching it for some reason. -DONE-
Have Pressing the Pause Button turn off the Pause Menu instead of needing to press Resume -DONE-
Refined not being able to jump and resume using the A Button at the same time -DONE-